Database view model (#5)
* Updated some stuff and added some tests

Im also kinda clueless as to how I should proceed. Datagrid does not support what I want to do or at least I didnt find it yet. It would require me to generate and compile classes on the fly

* Found possible solution online

* Fixed the Issue that prevented the DataGrid from correctly displaying the Data

* Data Binding now works and the table now changes depending on the supplied Query

* Moved some methods, added a new method, added a new ViewModel, renamed a VIewmodel

1. Moved the GetMySQLDatabases, GetMySQLTables functions to a new File.

2. Added a new GetTableFields Method

3. Added a new ViewModel named DatabaseStructureViewModel which is supposed to generate the treeview which lists the databases and tables etc.

* Added the required Binding and removed unnessesary stuff

* Rewritten the DatabaseTreeView method to properly reflect MVVM practices which also made it way faster hehehe

* Binding for the DatabaseTreeView Method, Fixed consistency issues regarding Namespaces

Implemented the Binding for the DatabaseTreeView( now called DatabaseModelConverter (not the final name)) Method

Fixed consistency issues regarding namsepaces

* DatabaseConnector now dynamic, Mysql autostarts

The Databaseconnector now automatically connects to the database selected in the treeview.

The XAMPP mysql server now automatically starts with the application if it is located in the root folder of the EXE. TODO: Let the user decide where it should be located. Requires some kind of config handler.