Pages (#6)
* Added Settings and implemented a settings page.

The next step is creating a Viewmodel for said page

* Added a janky solution to save the settings

* Added a search button, Fixed databinding on settings page, Settings now handeled through SettingsPageViewModel

Added a search button to the settingspage that allows the user to search for the executable of the mysql server.

Databinding for the settingspage is now done through XAML instead of the code behind. I will do this for all other pages soon.

* Added more options to the appliction and added them to the Settings page

Added the:
ServerLocation,
Username,
Password options

* Added first start window, Added binding to winheight,winwidth

* Changed type of some settings, added content to Firstart Window, Settings save on exit

* SelectedItemChanged crash fix

Fixed the crash that occured when selecting a diffrent item in the treeview by adding a reference of the parent to the TableModel. The application still crashed when the user selects a field though.

* fixed selectedItemChanged crash

* Renamed DataTableModel, DataFieldModel

* Import functionality which doesnt work